Add a script parallel run script for moodle's own PHPunit
This script uses Moodle's own copy of phpunit to perform a parallel run,
rather than relying on paratest and its seprate phpunit version.

It's currently very shonky, it blindly distributes testsuites between
threads round-robin style, and the output is all mixed together.
However, it does prove the concept and it could be useful as a way doing
a quick local run of a large set of tests.